Add Event to Temporal metrics abstraction and use it for mutable state stats
The family of metrics emitted in `emitMutableStateStatus` are each histograms that publish a single dimension of data. This commit adds an `Event` to Temporal's metrics abstraction that accepts a name identifying the event and, at runtime, an arbitrary dictionary can be given to the event with an arbitrary number of dimensions. This commit uses this abstraction to emit an event named `mutable_state_status` that has a dimension for each of the dimensions that are observed by the metrics emitted above it.